backend/WebWisdom/pyproject.toml
[tool.poetry]
name = "webwisdom"
version = "0.1.0"
description = ""
authors = ["Khaled Qasim Kraken <khaled.qasim183@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
python-multipart = "^0.0.7"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlalchemy = "^2.0.25"
httpx = "^0.26.0"
pytest = "^8.0.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0.post1"
argon2-cffi = "^23.1.0"
requests = "^2.31.0"
google-auth = "^2.28.0"
beautifulsoup4 = "^4.12.3"
aiohttp = "^3.9.3"
selenium = "^4.18.1"
webdriver-manager = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
